Add clip-path to header pseudo-element
Add a height to the
header, add a flex
display and make
its position.

Add an empty pseudo
element before the
header, give it an
absolute position and
style its top, left
right and bottom
properties so that
there is no distance
between it and the
header.
Add a clip-path property
to clip a region and give
it the shape.
Add a negative z-index
so that it stays at
the back of other
_children_ of the header.
Add a blueish background
color to it.