feat: Add projects' section
Add a main element and
add a projects-section
div inside it that
contains 6 projects.
Each project has a
project-screenshot and
project-details.
The project-details
contain the project
title, the project links
and the project description.

feat: Style the project's section

Add a padding to main.

Style projects-section
with a grid display,
define the columns
and add a gap.

Give .project a
min-height and set its
display to flex and
the direction to
column. Set the
flex-direction column
so that
project-screenshot can
be given a flex of
1 to take up more
space than the
project-details.

Style the title-and-links
with a display of flex
and justify its content.

Add padding to the
project-details.