Add footer section and content
Add a footer element. In the
footer, add a contact-details
div and a picture element.

Add the respective content
in contact details. Create
an images directory and add
the footer-image-large in it.
Add an img element in
in the picture
element and set the src
to be the newly added
footer image.