Style body and footer
Set a grid display on
the body and set the
rows accordingly.
Set a min-height so that
the body does not get
smaller than 100vh and
thus keeping the footer
always at the bottom
of the viewport.

Group the selectors
for the pseudo content
before the header and
the footer so that they
can share the same
background color.

Add a padding to the
footer and give it a
flex display. Set a
max-width for the
footer image.