Add header image and style
In the HTML file,
change the figure element
to a picture element and
remove the class.

Add a header image to the
images directory. Add
an img element in the
picture element set its
src to this header image.

Wrap this picture element
and the h1 in a
header-image-and-title
section
element.

In the CSS file, increase
the height and padding of
the header to get more
space.

Remove the .header-figure
selector and its styles.

Set the position of
the header-section-and-title
to be relative so that
the h1's position can be
set as absolute and then
positioned on top the
header image using
bottom and left. Use
the translate function
to adjust
horizontally and
vertically,
ensuring it is
placed at the bottom
center of the image
regardless of its
dimensions.

Set a max-width on the
header image.