Add about-me content and style
In the HTML file, add a
h2 element, a paragraph
and a div containing
social media links
and their respective
SVG icons.

In the CSS file, reduce
the height and width of
the about-me. Style
it with a flex display
and flex direction of
column so that the space
between its children can
be controlled with the
gap. Give it a
background color and
padding.

Add a left and right margin
to all children of
about-me.

Style the header social
media links with a flex
display and justify
content end so that
the children are
justified to the extreme
right.

Increase the width of
the header about me
svg icons.

Remove the grid display
and the rows definition
from the body as the
footer will stick to
the bottom with only the
body's min-height.

Change the header's
height to max-height and
increase the left and
right margin.