Style header texts.
Create a fonts directory
and add the Roboto and
Playfair display fonts.
Imports these fonts in the
CSS.

Style
.header-image-and-title h1
with the Playfair font family,
add a font color and
a font size. Increase the
width to a 100% to prevent
the text from wrapping
easily with bigger font.

Style the about-me h2
with the layfair font family,
and increase its font size.

Style the about-me paragraph
with the Roboto font family
and change its color.

Reduce the height and gap
of about-me.