Add icons to projects
Add github and
open-in-new
svg icons to the images
directory.

In the HTML file,
add these icons as the
src of the img elements
in each project.

In the CSS file, increase
the top and bottom paddings
of main. Style the
font family, increase the
font size and increase the
padding bottom of main's
h3.

Increase the width of
all images that are
main's descendants
(the svg icons).