Style main and projects
Increase the left, right
and bottom padding of
main. Increase the
padding-bottom of the
main h3.

Style the main h4 and
paragraph with the
respective
font-family and
increase their
font-sizes.

Add a near-subtle
box shadow effect
to .project.

Style project-details
with a flex display,
set its flex direction
to column and add a gap
to control the space
between title-and-links
and description.