Link header SVGs with img
In the HTML file, Add
linkedin and twitter(x)
svg icons in the images
directory

Remove the inline
embedding of the svg
icons
in the header. Use
the img element to
link them instead.

Change the
project-screenshot
to div since I
do not intend to
include any
screenshots for now.

In the CSS file,
change the selector
to target header-image
by its class name.

Also change the
selector targeting
the icons on the
header to target
img elements
since the icons
have been changed
from inline svg
to
the img element
(linked).