Style footer
In the CSS,
add a gap to footer
to control the space
between
contact-details and
footer-image.

Style contact-details
with a flex display,
a flex-direction
of column and a gap.

Style contact-details
h4 and paragraph
with a font family,
color and increase
their font-sizes.

Style the footer-image
with a max-width of
55vw to make it
more responsive.

In the HTML file,
correct a
typographical
error in the
footer
contact-details.