Add icons and style
In the HTML page in the
.contact-details,
add inline SVGs for
the phone and the
email icons.
Add a fill color and
a width to these SVG
icons via inline
styling.

In .social-media-links
add SVG icons for
github, linkedin and
x(twitter).

In the CSS file,
increase the font-size
of contact-details h4.

Group the selectors
of contact-details
paragraph,
phone-container and
email-container so tht
they can be styled with
the same font family,
size,
weight and color.

Style phone-container
and email-container
with a flex display
and a gap. Align
the phone-number
and email-number
to the center
so that they line up
vertically with
their icons.

Style social-media-links
with a flex display and
gap so that the space
between the between
the icons can be
controlled.

Increase the width of
the social media
icons and decrease
the width of the
footer image.