Set media query for tablets
Set media query for
screens larger than
600px and smaller
than 768px.

In the media query,
add a different
clip-path for the
pseudo element before
the header so that it
conforms to the
header about be section
having a larger width.

Reduce the padding-top
of the header.

Set the header-image
position to be
absolute so that
it is removed from
the normal document
flow.

Set the min-width
of about-me to
be 75vw so that it
is wider.

Set the display of
the header social
media links to none
as they won't be
needed in the
tablet display.