Style header for tablet display
Remove the set height and
width
styles about-me
and increase the padding
padding top and bottom
so that it can be long
enough to cover the
bottom of the
header::before pseudo
element as regards the
layout.

In the media query,
increase the max-width
so that this layout
can apply to wider
tablets.

In the media query,
increase the
padding-top of the
header.
Set the header image
position to relative
and set it left by
10% so that its right
edge is overlapping
the about-me.

Reduce the padding
of about me.
Increase the
font size of the
about-me paragraph.