Style projects and footer
In the media query,
reduce the size of the
column tracks for the
projects-section
so that the projects can
line up in twos.

Reduce the min-height of
project because they
were too long.

Reduce the gap of footer.

Have the footer-image
to align itself center
so that it align better
with contact-details.