Set media query for mobiles
First remove the max-height
from header so that it
can be as high as its
children. Reduce its
padding top so that
there is no much space
at the top.
Change the clip path
values for the
pseudo-element
before header so
that the
header image and
about-me
can align with
the pseudo-element
and fit design
requirements.

Set the media query
to apply for all
devices that have a
viewport width
of 600px and below.

In the media query,
set new clip-path
values for the
pseudo-element
before the header
to fit with
design requirements.

Set the header's flex
direction to column
to that the header
image and the about-me
can line up as
required. Decrease
its padding-top.

Set a larger max-width
for the header-image.

Set a padding for
about-me and remove
its box shadow.

Align the about-me
h2 to the center.

Add a width to the
about-me paragraph
so that it gets
wider and oushes
about-me close to
the edges of the
page as required.