Style main for mobiles
Change the selector
targeting the image
icons in project
to more precise
selectors
(.links img)

Reduce the
main h3
padding bottom.

In the styles
outside media queries
and in the media query
for tablets,change the
min-height of
project to height
so that it does not
get larger or smaller
except
styled on purpose.

In the media query for
mobiles, justify the
social-media-links-header
(icons)
content to the center
and increase the width of
the images.

Remove the padding-top
from main and reduce its
left and right padding.

Align the main h3 to the
center and increase
the size of the
.links img.