Style footer for mobiles
First remove the padding
left and right
from about-me and the
padding from
about me h3, paragraph
and div because
excess space cause
the page to overflow
on the left and right.

Set a column
flex direction
to the footer so that
contact-details and
footer-image can line
up vertically.
Remove the footer's
padding-bottom so that
the footer-image can
stay at the bottom of
the footer.

Align the footer-image
container to the center
and give the image
a max-width of 100vw
so that it never gets
wider than 100vw.
The image now covers
the bottom of the footer.