Adjust footer padding for mobiles
In the CSS file, first
add a padding-top
to about-me h2.

Reduce the
left and right padding
for footer so that the
text in contact-details
can get closer to right
and left edges of the
page.

Remove the svg icons of
phone and email containers
from display with
display:none.

Group the selectors
of footer-image
and social-media-links
so that they both
can be set to
align-self to center.

Increase the width of
the social media icons.

In the HTML, update
the phone number
format.