Implement responsive images
Resize the header and footer
images to 1000px, 800px and
400px widths and convert
them to WEBP format.

Delete the older header and
footer images.

In the HTML, for the header
and footer images, implement
responsive images by using
`Resolution switching:
Different sizes`
(Look it up on MDN).

In the CSS file, add
a comment.